requirements.txt
eth-brownie>=1.0.0b6
